backend/requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.3
click-plugins==1.1.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
h5py==3.2.1
iniconfig==1.0.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
Keras==2.4.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lief==0.10.1
lightgbm==3.0.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.5.0
mpmath==1.2.1
multidict==4.7.6
names==0.3.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
networkx==2.5
notebook==6.1.4
numpy==1.19.2
opencv-contrib-python==4.5.1.48
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
powerline-gitstatus==1.3.1
powerline-status==2.7
prettytable==2.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
py==1.9.0
pycparser==2.20
Pygments==2.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.0.1
pytest-httpserver==0.3.5
python-dateutil==2.8.1
pytz==2021.1
pyzmq==19.0.2
scapy==2.4.4
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
Send2Trash==1.5.0
shodan==1.24.0
six==1.15.0
sklearn==0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
virtualenv==20.2.1
vt-py==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xgboost==1.3.3
xlrd==1.2.0
XlsxWriter==1.3.7
yarl==1.5.1